Implement drawing chess board
This is pretty clever - we have a ChessFrame class, that inherits from
JFrame, and handle displaying all the things related with actual game.
Then we can create another class that inherits form JFrame but displays
something different, like menu.
If we would like to switch the Frames, we can just class setVisible()
and voila.