Implement Working Piece Movement
Previously, we could only move the graphical representation of the
piece. Because of that, we couldn't get the available moves for a
previously moved piece. When we clicked on the square where that piece
used to be, it still showed the moves.

Now, we can move the actual piece. Other pieces disappear when we place
another piece on them, and we can't place a piece on top of another
piece of the same color—otherwise, it would disappear, as if it captured
itself.

It also required to hide available moves after mouse releasing,
otherwise the move indicator and piece overlapped each other and it
causes bugs with displaying.

Overall, developing this chess project has taught me a lot so far.