Start using Move object
I used to pass all 4 variables for any move. This was quite cumbersome,
but now I have only one. Mainly that was the reason for this change.

Additionally, I changed the way we use ChessFrame for making moves.
Previously, it has makeMove() function, which just called ChessPanel
makeMove(). making interface through ChessFrame just repeated the code,
so I removed this and created a function, which returns reference to
ChessPanel object. Then we execute makeMove on this.